# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Data Template Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List templates with optional filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: tags
          in: query
          schema:
            type: ["string", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Paginated template list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResult"

    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new template (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
...
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a single template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateDetail"
        "404":
          description: Not found

    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"

    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /templates/{id}/versions:
    get:
      operationId: listTemplateVersions
      tags: [templates]
      summary: Get all versions of a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemplateVersion"
  /templates/{id}/download:
    post:
      operationId: downloadTemplate
      tags: [templates]
      summary: Record a template download
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateContent"
...
components:
  schemas:
...
    Template:
      type: object
      required: [id, title, description, categoryId, categoryName, authorId, authorName, version, status, downloads, averageRating, ratingCount, isFavorited, tags, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        version:
          type: string
        status:
          type: string
        downloads:
          type: integer
        averageRating:
          type: number
        ratingCount:
          type: integer
        isFavorited:
          type: boolean
        tags:
          type: array
          items:
            type: string
        language:
          type: ["string", "null"]
        complexity:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TemplateDetail:
      type: object
      required: [id, title, description, categoryId, categoryName, authorId, authorName, version, status, downloads, averageRating, ratingCount, isFavorited, tags, code, documentation, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        version:
          type: string
        status:
          type: string
        downloads:
          type: integer
        averageRating:
          type: number
        ratingCount:
          type: integer
        isFavorited:
          type: boolean
        tags:
          type: array
          items:
            type: string
        language:
          type: ["string", "null"]
        complexity:
          type: ["string", "null"]
        code:
          type: string
        documentation:
          type: string
        usageExample:
          type: ["string", "null"]
        prerequisites:
          type: ["string", "null"]
        changelog:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TemplateContent:
      type: object
      required: [code, filename]
      properties:
        code:
          type: string
        filename:
          type: string
    TemplateListResult:
      type: object
      required: [templates, total, page, limit]
      properties:
        templates:
          type: array
          items:
            $ref: "#/components/schemas/Template"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TemplateInput:
      type: object
      required: [title, description, categoryId, code, tags, version]
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        code:
          type: string
        documentation:
          type: string
        usageExample:
          type: string
        prerequisites:
          type: string
        tags:
          type: array
          items:
            type: string
        version:
          type: string
        language:
          type: string
        complexity:
          type: string
    TemplateUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        code:
          type: string
        documentation:
          type: string
        tags:
          type: array
          items:
            type: string
        version:
          type: string
        status:
          type: string
        language:
          type: string
        complexity:
          type: string
    TemplateVersion:
      type: object
      required: [id, templateId, version, changelog, createdAt, authorName]
      properties:
        id:
          type: integer
        templateId:
          type: integer
        version:
          type: string
        changelog:
          type: string
        code:
          type: string
        createdAt:
          type: string
        authorName:
          type: string
...