# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Template Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateInput {
  title: string;
  description: string;
  categoryId: number;
  code: string;
  documentation?: string;
  usageExample?: string;
  prerequisites?: string;
  tags: string[];
  version: string;
  language?: string;
  complexity?: string;
}