# lib/api-zod/src/generated/types/templateVersion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Data Template Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface TemplateVersion {
  id: number;
  templateId: number;
  version: string;
  changelog: string;
  code?: string;
  createdAt: string;
  authorName: string;
}